+++ cmd/hidden_lake/traffic/internal/database/settings.go
@@ -9,6 +9,7 @@ type sSettings struct {
 	FPath             string
 	FNetworkKey       string
 	FWorkSizeBits     uint64
+	FHashesWindow     uint64
 	FMessagesCapacity uint64
 }
 
@@ -17,6 +18,7 @@ func NewSettings(pSett *SSettings) ISettings {
 		FPath:             pSett.FPath,
 		FNetworkKey:       pSett.FNetworkKey,
 		FWorkSizeBits:     pSett.FWorkSizeBits,
+		FHashesWindow:     pSett.FHashesWindow,
 		FMessagesCapacity: pSett.FMessagesCapacity,
 	}).mustNotNull()
 }
@@ -24,6 +26,9 @@ func NewSettings(pSett *SSettings) ISettings {
 func (p *sSettings) mustNotNull() ISettings {
 	if p.FPath == "" {
 		panic(`p.FPath == ""`)
+	}
+	if p.FHashesWindow == 0 {
+		panic("p.FHashesWindow == 0")
 	}
 	if p.FMessagesCapacity == 0 {
 		panic("p.FMessagesCapacity == 0")
@@ -33,6 +38,10 @@ func (p *sSettings) mustNotNull() ISettings {
 
 func (p *sSettings) GetPath() string {
 	return p.FPath
+}
+
+func (p *sSettings) GetHashesWindow() uint64 {
+	return p.FHashesWindow
 }
 
 func (s *sSettings) GetMessagesCapacity() uint64 {