+++ cmd/hidden_lake/traffic/internal/handler/handler_test.go
@@ -42,6 +42,7 @@ func testAllRun(addr, addrNode string) (*http.Server, conn_keeper.IConnKeeper, d
 			FPath:             fmt.Sprintf(databaseTemplate, addr),
 			FNetworkKey:       testutils.TCNetworkKey,
 			FWorkSizeBits:     testutils.TCWorkSize,
+			FHashesWindow:     testutils.TCCapacity,
 			FMessagesCapacity: testutils.TCCapacity,
 		}),
 	)